[chore] Refactor the code and implement blueprints
